+++ multicam/core/imaging/__init__.py
@@ -1,3 +1,9 @@
 from .compositor import Compositor
+from .dsp import DspConfig, DspPipeline, DspPipelineStore
 
-__all__ = ["Compositor"]
+__all__ = [
+    "Compositor",
+    "DspConfig",
+    "DspPipeline",
+    "DspPipelineStore",
+]