+++ multicam/core/services/__init__.py
@@ -3,6 +3,7 @@ from .camera_profiles import CameraProfileStore
 from .alignment import AlignmentService, FrozenFrameInfo
 from .camera_orientation import CameraOrientationStore
 from .focus import FocusSample, FocusService
+from .dsp import DspService, DspStatus
 
 __all__ = [
     "AlignmentService",
@@ -12,4 +13,6 @@ __all__ = [
     "FocusSample",
     "FocusService",
     "LiveViewService",
+    "DspService",
+    "DspStatus",
 ]